+++ src/main/java/ReviewRatingUseCase/DocumentReviewConverter.java
@@ -1,4 +1,4 @@
-package Interactors;
+package ReviewRatingUseCase;
 
 import Entities.*;
